+++ frontend/app/me/dashboard/components/data-table/columns.tsx
@@ -10,7 +10,6 @@ import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
-  DropdownMenuSeparator,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 import { Project, ProjectStatus } from "@/types/models/user";
